get_code_executor as autogen_util
