+++ openai_server/autogen_agents.py
@@ -5,52 +5,10 @@ from openai_server.agent_utils import in_pycharm, set_python_path
 from openai_server.autogen_utils import terminate_message_func
 
 
-def get_execution_agent(
-        autogen_run_code_in_docker,
-        autogen_timeout,
-        agent_system_site_packages,
+def get_code_execution_agent(
+        executor,
         autogen_max_consecutive_auto_reply,
-        autogen_code_restrictions_level,
-        agent_venv_dir,
-        temp_dir
         ):
-    if autogen_run_code_in_docker:
-        from autogen.coding import DockerCommandLineCodeExecutor
-        # Create a Docker command line code executor.
-        executor = DockerCommandLineCodeExecutor(
-            image="python:3.10-slim-bullseye",
-            timeout=autogen_timeout,  # Timeout for each code execution in seconds.
-            work_dir=temp_dir,  # Use the temporary directory to store the code files.
-        )
-    else:
-        set_python_path()
-        from autogen.code_utils import create_virtual_env
-        if agent_venv_dir is None:
-            username = str(uuid.uuid4())
-            agent_venv_dir = ".venv_%s" % username
-        env_args = dict(system_site_packages=agent_system_site_packages,
-                        with_pip=True,
-                        symlinks=True)
-        if not in_pycharm():
-            virtual_env_context = create_virtual_env(agent_venv_dir, **env_args)
-        else:
-            print("in PyCharm, can't use virtualenv, so we use the system python", file=sys.stderr)
-            virtual_env_context = None
-        # work_dir = ".workdir_%s" % username
-        # PythonLoader(name='code', ))
-
-        # Create a local command line code executor.
-        if autogen_code_restrictions_level >= 2:
-            from autogen_utils import H2OLocalCommandLineCodeExecutor
-        else:
-            from autogen.coding.local_commandline_code_executor import \
-                LocalCommandLineCodeExecutor as H2OLocalCommandLineCodeExecutor
-        executor = H2OLocalCommandLineCodeExecutor(
-            timeout=autogen_timeout,  # Timeout for each code execution in seconds.
-            virtual_env_context=virtual_env_context,
-            work_dir=temp_dir,  # Use the temporary directory to store the code files.
-        )
-
     # Create an agent with code executor configuration.
     from openai_server.autogen_utils import H2OConversableAgent
     code_executor_agent = H2OConversableAgent(
@@ -59,69 +17,6 @@ def get_execution_agent(
         code_execution_config={"executor": executor},  # Use the local command line code executor.
         human_input_mode="NEVER",  # Always take human input for this agent for safety.
         is_termination_msg=terminate_message_func,
-        max_consecutive_auto_reply=autogen_max_consecutive_auto_reply,
-    )
-    return code_executor_agent, executor
-
-
-def get_code_executor(
-    temp_dir,
-    agent_system_site_packages,
-    agent_venv_dir: str,
-    autogen_run_code_in_docker: bool = False,
-    autogen_timeout: int = 60,
-    autogen_code_restrictions_level: int = 2,
-):
-    from autogen.coding import DockerCommandLineCodeExecutor
-    if autogen_run_code_in_docker:
-        from autogen.coding import DockerCommandLineCodeExecutor
-        # Create a Docker command line code executor.
-        executor = DockerCommandLineCodeExecutor(
-            image="python:3.10-slim-bullseye",
-            timeout=autogen_timeout,  # Timeout for each code execution in seconds.
-            work_dir=temp_dir,  # Use the temporary directory to store the code files.
-        )
-    else:
-        set_python_path()
-        from autogen.code_utils import create_virtual_env
-        if agent_venv_dir is None:
-            username = str(uuid.uuid4())
-            agent_venv_dir = ".venv_%s" % username
-        env_args = dict(system_site_packages=agent_system_site_packages,
-                        with_pip=True,
-                        symlinks=True)
-        if not in_pycharm():
-            virtual_env_context = create_virtual_env(agent_venv_dir, **env_args)
-        else:
-            print("in PyCharm, can't use virtualenv, so we use the system python", file=sys.stderr)
-            virtual_env_context = None
-        # work_dir = ".workdir_%s" % username
-        # PythonLoader(name='code', ))
-
-        # Create a local command line code executor.
-        if autogen_code_restrictions_level >= 2:
-            from autogen_utils import H2OLocalCommandLineCodeExecutor
-        else:
-            from autogen.coding.local_commandline_code_executor import \
-                LocalCommandLineCodeExecutor as H2OLocalCommandLineCodeExecutor
-        executor = H2OLocalCommandLineCodeExecutor(
-            timeout=autogen_timeout,  # Timeout for each code execution in seconds.
-            virtual_env_context=virtual_env_context,
-            work_dir=temp_dir,  # Use the temporary directory to store the code files.
-        )
-    return executor
-
-# TODO: Reuse get_execution_agent ?
-def get_code_executor_agent(
-        executor,
-        autogen_max_consecutive_auto_reply: int = 1,
-        ):
-    from openai_server.autogen_utils import H2OConversableAgent
-    code_executor_agent = H2OConversableAgent(
-        name="code_executor_agent",
-        llm_config=False, 
-        code_execution_config={"executor": executor},
-        human_input_mode="NEVER", 
         max_consecutive_auto_reply=autogen_max_consecutive_auto_reply,
     )
     return code_executor_agent
@@ -211,7 +106,7 @@ def get_code_group_chat_manager(
         llm_config=llm_config,
         autogen_max_consecutive_auto_reply=autogen_max_consecutive_auto_reply,
     )
-    code_executor_agent = get_code_executor_agent(
+    code_executor_agent = get_code_execution_agent(
         executor=executor,
         autogen_max_consecutive_auto_reply=autogen_max_consecutive_auto_reply,
     )
